pyproject.toml
[project]
name = "telegram-bot-base"
version = "0.1.2"
description = "Base reutilizable para bots de Telegram con logging y estructura común."
authors = [{ name = "Pablo Martin", email = "pablo.martin95@hotmail.com" }]
dependencies = [
    "python-telegram-bot>=20.0",
    "python-decouple>=3.8",
    "pandas>=1.0.0",
    "python-dotenv>=1.0.0"
]
readme = "README.md"
requires-python = ">=3.9"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
